:truck: chore: rename Newsletter ShareForm
